in release version the following compiler Warnings are disabled:
4275;
4251;
4099;

the compiler is complaining about that objects are used over binary boundaries what is considered to be dangerous. but since we are compiling all project together with the same compiler flags it is ok. the main idea to put them on the ignore list is to not confuse our costumers when they want to make some changes. so that they only get the messages that are important for them.